chore: move 'borrowed_datum' under its own module.
